+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+Enhancements:
+* [#1557][]: Add a `Context` option and context-aware `*Ctx` logging methods (`InfoCtx`, `ErrorCtx`, etc. on `Logger`; `InfowCtx`, `InfofCtx`, etc. on `SugaredLogger`) that extract fields from a `context.Context`.
+
+[#1557]: https://github.com/uber-go/zap/pull/1557
+
 ## 1.28.0 (27 Apr 2026)
 Enhancements:
 * [#1534][]: Add `zapcore.CheckPreWriteHook` and `CheckedEntry.Before` method for transforming entries before they are written to any Cores.